# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "REST API members"
  },
  "basePath": "/api/v1",
  "paths": {
    "/members": {
      "get": {
        "tags": [
          "members"
        ],
        "summary": "Read",
        "responses": {
          "200": {
            "description": "Read"
          }
        }
      },
      "post": {
        "tags": [
          "members"
        ],
        "summary": "Create",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/member"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Created"
          }
        }
      }
    },
    "/members/{index}": {
      "patch": {
        "tags": [
          "members"
        ],
        "summary": "Update",
        "parameters": [
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "number",
            "default": 0
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/member"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated"
          }
        }
      },
      "delete": {
        "tags": [
          "members"
        ],
        "summary": "Delete",
        "parameters": [
          {
            "in": "path",
            "name": "index",
            "required": true,
            "type": "number",
            "default": 0
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted"
          }
        }
      }
    }
  },
  "definitions": {
    "member": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string",
          "example": "홍길선"
        },
        "age": {
          "type": "string",
          "example": "38"
        }
      }
    }
  }
}